openaiv1/requirements.txt
openai
dotenv
uvicorn
fastapi
pydantic
mysql-connector-python